added fake type for docs
